+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.3.3
 
 - Add `HandleClient` to retry request on error
-- Throw exceptions that are `ClientException`s when catching errors in convertors and interceptors
+- Throw exceptions when catching errors in convertors and interceptors
 
 ## 0.3.2+2